code/shared/location/gps_sensor/pyproject.toml
[tool.poetry]
name = "gps-sensor"
version = "0.1.0"
description = ""
authors = [
    "Shahar Tal <stalmail10@gmail.com>",
    "Or Peretz <or3325@gmail.com>"
]

[tool.poetry.dependencies]
python = "^3.11"
gpsd-py3 = "^0.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
